backend/requirements.txt
Django==4.2.*
djangorestframework==3.14.*
psycopg2-binary==2.9.*
python-decouple==3.8.*
celery==5.3.*
redis==5.0.*
boto3==1.34.*
django-cors-headers==4.3.*
stripe==8.0.*
django-celery-beat==2.5.*
